feat:登录、注册、忘记密码界面排版以及功能实现

+++ client/src/api/index.ts
@@ -20,7 +20,7 @@ instance.interceptors.request.use((config: AxiosRequestConfig) => {
 instance.interceptors.response.use((response: any) => {
   if (response.data.code === 200) {
     return Promise.resolve(response.data.data)
-  } else {
+  }else {
     return Promise.reject(response.data)
   }
 }, (error) => {
@@ -37,13 +37,17 @@ function http (url, data) {
       resolve(res)
     }).catch(error => {
       console.log('报错了: ', error)
-      ElMessage({ message: error.data || error.message, type: 'error' })
-      if ([401].includes(error.code)) {
-        localStorage.clear()
-        useUserStore().updateUserInfo(null)
-        router.push({
-          path: '/login'
-        })
+      if ([201].includes(error.code)) {
+        reject(error)
+      } else {
+        ElMessage({ message: error.data || error.message, type: 'error' })
+        if ([401].includes(error.code)) {
+          localStorage.clear()
+          useUserStore().updateUserInfo(null)
+          router.push({
+            path: '/login'
+          })
+        }
       }
     })
   })
